Moved the header to a separate folder
