+++ my-vue-app/src/App.vue
@@ -1,5 +1,7 @@
 <script setup>
-import Header from './components/Header.vue';
+//import Header from './components/Header/Header.vue';
+
+import Header from './components/Header/Header.vue';
 import MainBanner from './components/MainBanner.vue';
 </script>
